restructure route
