+++ apps/editor.planx.uk/src/routes/_authenticated/app/$team/submissions/index.tsx
@@ -1,9 +1,24 @@
 import { createFileRoute } from "@tanstack/react-router";
+import { hasFeatureFlag } from "lib/featureFlags";
+import SubmissionDetailModal from "pages/FlowEditor/components/Submissions/components/SubmissionDetailModal";
+import Submissions from "pages/FlowEditor/components/Submissions/Submissions";
+import SubmissionsGrouped from "pages/FlowEditor/components/Submissions/SubmissionsGrouped";
 
 export const Route = createFileRoute("/_authenticated/app/$team/submissions/")({
   component: RouteComponent,
 });
 
 function RouteComponent() {
-  return null;
+  const { detail } = Route.useSearch();
+
+  const SubmissionsWrapper = hasFeatureFlag("GROUPED_SUBMISSIONS")
+    ? SubmissionsGrouped
+    : Submissions;
+
+  return (
+    <>
+      <SubmissionsWrapper />
+      {detail && <SubmissionDetailModal sessionId={detail} />}
+    </>
+  );
 }
