--- apps/editor.planx.uk/src/routes/_authenticated/app/$team/view-submission/$sessionId.tsx
@@ -1,17 +1,0 @@
-import { createFileRoute } from "@tanstack/react-router";
-import SubmissionHTML from "pages/FlowEditor/components/Submissions/components/SubmissionHTML";
-import React from "react";
-
-export const Route = createFileRoute(
-  "/_authenticated/app/$team/view-submission/$sessionId",
-)({
-  loader: async ({ params }) => ({
-    sessionId: params.sessionId,
-  }),
-  component: RouteComponent,
-});
-
-function RouteComponent() {
-  const { sessionId } = Route.useLoaderData();
-  return <SubmissionHTML sessionId={sessionId} />;
-}
